Combine generator and parser into new translator module
Also update related imports
