+++ rmgpy/molecule/molecule.py
@@ -56,8 +56,7 @@ from .atomtype import AtomType, atomTypes, getAtomType, AtomTypeError
 import rmgpy.constants as constants
 import rmgpy.molecule.element as elements
 import rmgpy.molecule.converter as converter
-import rmgpy.molecule.parser as parser
-import rmgpy.molecule.generator as generator
+import rmgpy.molecule.translator as translator
 import rmgpy.molecule.resonance as resonance
 from .kekulize import kekulize
 from .adjlist import Saturator
@@ -1319,21 +1318,21 @@ class Molecule(Graph):
         """
         Convert an InChI string `inchistr` to a molecular structure.
         """
-        parser.fromInChI(self, inchistr, backend)
+        translator.fromInChI(self, inchistr, backend)
         return self
 
     def fromAugmentedInChI(self, aug_inchi):
         """
         Convert an Augmented InChI string `aug_inchi` to a molecular structure.
         """
-        parser.fromAugmentedInChI(self, aug_inchi)
+        translator.fromAugmentedInChI(self, aug_inchi)
         return self
 
     def fromSMILES(self, smilesstr, backend='try-all'):
         """
         Convert a SMILES string `smilesstr` to a molecular structure.
         """
-        parser.fromSMILES(self, smilesstr, backend)
+        translator.fromSMILES(self, smilesstr, backend)
         return self
         
     def fromSMARTS(self, smartsstr):
@@ -1342,7 +1341,7 @@ class Molecule(Graph):
         `RDKit <http://rdkit.org/>`_ to perform the conversion.
         This Kekulizes everything, removing all aromatic atom types.
         """
-        parser.fromSMARTS(self, smartsstr)
+        translator.fromSMARTS(self, smartsstr)
         return self
 
     def fromAdjacencyList(self, adjlist, saturateH=False):
@@ -1414,7 +1413,7 @@ class Molecule(Graph):
         Convert a molecular structure to an InChI string. Uses
         `OpenBabel <http://openbabel.org/>`_ to perform the conversion.
         """
-        return generator.toInChI(self)            
+        return translator.toInChI(self)
         
     def toAugmentedInChI(self):
         """
@@ -1423,7 +1422,7 @@ class Molecule(Graph):
         
         Separate layer with a forward slash character.
         """
-        return generator.toAugmentedInChI(self)
+        return translator.toAugmentedInChI(self)
         
     
     def toInChIKey(self):
@@ -1439,7 +1438,7 @@ class Molecule(Graph):
         Removes check-sum dash (-) and character so that only 
         the 14 + 9 characters remain.
         """
-        return generator.toInChIKey(self)
+        return translator.toInChIKey(self)
     
     def toAugmentedInChIKey(self):
         """
@@ -1449,7 +1448,7 @@ class Molecule(Graph):
         Simply append the multiplicity string, do not separate by a
         character like forward slash.
         """
-        return generator.toAugmentedInChIKey(self)
+        return translator.toAugmentedInChIKey(self)
     
 
     def toSMARTS(self):
@@ -1458,7 +1457,7 @@ class Molecule(Graph):
         `RDKit <http://rdkit.org/>`_ to perform the conversion.
         Perceives aromaticity and removes Hydrogen atoms.
         """
-        return generator.toSMARTS(self)
+        return translator.toSMARTS(self)
     
     
     def toSMILES(self):
@@ -1475,7 +1474,7 @@ class Molecule(Graph):
         and removes Hydrogen atoms.
         """
         
-        return generator.toSMILES(self)
+        return translator.toSMILES(self)
 
     def toRDKitMol(self, *args, **kwargs):
         """
