+++ rmgpy/molecule/translator.py
@@ -28,13 +28,15 @@
 #                                                                             #
 ###############################################################################
 
-# global imports
+"""
+This module provides methods for translating to and from common molecule
+representation formats, e.g. SMILES, InChI, SMARTS.
+"""
+
 
 import cython
-import logging
 import itertools
-
-# local imports
+import logging
 
 # Assume that OB is not installed by default
 INSTALLED_BACKENDS = {
@@ -44,12 +46,12 @@ INSTALLED_BACKENDS = {
 try:
     import openbabel
     INSTALLED_BACKENDS['OB'] = True
-except :
+except:
     pass
-
 from rdkit import Chem
 
-from rmgpy.molecule.converter import fromRDKitMol, fromOBMol
+from .molecule import Atom
+from rmgpy.molecule.converter import toRDKitMol, fromRDKitMol, toOBMol, fromOBMol
 
 import rmgpy.molecule.inchi as inchiutil
 import rmgpy.molecule.util as util
@@ -57,42 +59,255 @@ import rmgpy.molecule.util as util
 # constants
 
 BACKENDS = [
-            'rdkit',
-            ]
+    'rdkit',
+]
 
 if INSTALLED_BACKENDS['OB']:
-    BACKENDS.insert(0,'openbabel')
+    BACKENDS.insert(0, 'openbabel')
 
 INCHI_LOOKUPS = {
-            'H': '[H]',#RDkit was improperly handling the Hydrogen radical from InChI
-            'He': '[He]',
-        }
+    'H': '[H]',  # RDkit was improperly handling the Hydrogen radical from InChI
+    'He': '[He]',
+}
 SMILES_LOOKUPS = {
     '[He]':  # RDKit improperly handles helium and returns it in a triplet state
-            """
-            He
-            multiplicity 1
-            1 He u0 p1
-            """,
+        """
+        He
+        multiplicity 1
+        1 He u0 p1
+        """,
     '[Ar]':  # RDKit improperly handles argon
-            """
-            Ar
-            multiplicity 1
-            1 Ar u0 p4
-            """,
+        """
+        Ar
+        multiplicity 1
+        1 Ar u0 p4
+        """,
     '[C]':  # We'd return the quintuplet without this
-            """
-            multiplicity 3
-            1 C u2 p1 c0
-            """,
+        """
+        multiplicity 3
+        1 C u2 p1 c0
+        """,
     '[CH]':  # We'd return the quartet without this
-            """
-            multiplicity 2
-            1 C u1 p1 c0 {2,S}
-            2 H u0 p0 c0 {1,S}
-            """,
+        """
+        multiplicity 2
+        1 C u1 p1 c0 {2,S}
+        2 H u0 p0 c0 {1,S}
+        """,
+}
 
-}     
+#: This dictionary is used to shortcut lookups of a molecule's SMILES string from its chemical formula.
+_known_smiles_molecules = {
+    'N2': 'N#N',
+    'CH4': 'C',
+    'CH2O': 'C=O',
+    'H2O': 'O',
+    'C2H6': 'CC',
+    'H2': '[H][H]',
+    'H2O2': 'OO',
+    'C3H8': 'CCC',
+    'Ar': '[Ar]',
+    'He': '[He]',
+    'CH4O': 'CO',
+    'CO2': 'O=C=O',
+    'CO': '[C-]#[O+]',
+    'O2': 'O=O',
+    'C': '[C]',  # for this to be in the "molecule" list it must be singlet with 2 lone pairs
+    'H2S': 'S',
+    'N2O': 'N#[N+][O-]',
+    'NH3': 'N',
+    'O3': '[O-][O+]=O',
+    'Cl2': '[Cl][Cl]',
+    'ClH': 'Cl',
+    'I2': '[I][I]',
+    'HI': 'I',
+}
+
+_known_smiles_radicals = {
+    'CH3': '[CH3]',
+    'HO': '[OH]',
+    'C2H5': 'C[CH2]',
+    'O': '[O]',
+    'S': '[S]',
+    'N': '[N]',
+    'HO2': '[O]O',
+    'CH': '[CH]',
+    'CH2': '[CH2]',
+    'H': '[H]',
+    'C': '[C]',  # this, in the radical list, could be triplet or quintet.
+    'O2': '[O][O]',
+    'S2': '[S][S]',
+    'OS': '[S][O]',
+    'HS': '[SH]',
+    'H2N': '[NH2]',
+    'HN': '[NH]',
+    'NO': '[N]=O',
+    'NO2': 'N(=O)[O]',
+    'Cl': '[Cl]',
+    'I': '[I]',
+}
+
+
+def toInChI(mol):
+    """
+    Convert a molecular structure to an InChI string. Uses
+    `RDKit <http://rdkit.org/>`_ to perform the conversion.
+    Perceives aromaticity.
+
+    or
+
+    Convert a molecular structure to an InChI string. Uses
+    `OpenBabel <http://openbabel.org/>`_ to perform the conversion.
+    """
+    try:
+        if not Chem.inchi.INCHI_AVAILABLE:
+            return "RDKitInstalledWithoutInChI"
+        rdkitmol = toRDKitMol(mol)
+        return Chem.inchi.MolToInchi(rdkitmol, options='-SNon')
+    except:
+        pass
+
+    obmol = toOBMol(mol)
+    obConversion = openbabel.OBConversion()
+    obConversion.SetOutFormat('inchi')
+    obConversion.SetOptions('w', openbabel.OBConversion.OUTOPTIONS)
+    return obConversion.WriteString(obmol).strip()
+
+
+def toAugmentedInChI(mol):
+    """
+    This function generates the augmented InChI canonical identifier, and that allows for the differentiation
+    between structures with spin states and multiple unpaired electrons.
+
+    Two additional layers are added to the InChI:
+    - unpaired electrons layer: the position of the unpaired electrons in the molecule
+
+    """
+
+    cython.declare(
+        inchi=str,
+        ulayer=str,
+        aug_inchi=str,
+    )
+    inchi = toInChI(mol)
+
+    ulayer, player = inchiutil.create_augmented_layers(mol)
+
+    aug_inchi = inchiutil.compose_aug_inchi(inchi, ulayer, player)
+
+    return aug_inchi
+
+
+def toInChIKey(mol):
+    """
+    Convert a molecular structure to an InChI Key string. Uses
+    `OpenBabel <http://openbabel.org/>`_ to perform the conversion.
+
+    or
+
+    Convert a molecular structure to an InChI Key string. Uses
+    `RDKit <http://rdkit.org/>`_ to perform the conversion.
+
+    Removes check-sum dash (-) and character so that only
+    the 14 + 9 characters remain.
+    """
+    try:
+        if not Chem.inchi.INCHI_AVAILABLE:
+            return "RDKitInstalledWithoutInChI"
+        inchi = toInChI(mol)
+        return Chem.inchi.InchiToInchiKey(inchi)[:-2]
+    except:
+        pass
+
+    # for atom in mol.vertices:
+    #           if atom.isNitrogen():
+    obmol = toOBMol(mol)
+    obConversion = openbabel.OBConversion()
+    obConversion.SetOutFormat('inchi')
+    obConversion.SetOptions('w', openbabel.OBConversion.OUTOPTIONS)
+    obConversion.SetOptions('K', openbabel.OBConversion.OUTOPTIONS)
+    return obConversion.WriteString(obmol).strip()[:-2]
+
+
+def toAugmentedInChIKey(mol):
+    """
+    Adds additional layers to the InChIKey,
+    generating the "augmented" InChIKey.
+    """
+
+    cython.declare(
+        key=str,
+        ulayer=str
+    )
+
+    key = toInChIKey(mol)
+
+    ulayer, player = inchiutil.create_augmented_layers(mol)
+
+    return inchiutil.compose_aug_inchi_key(key, ulayer, player)
+
+
+def toSMARTS(mol):
+    """
+    Convert a molecular structure to an SMARTS string. Uses
+    `RDKit <http://rdkit.org/>`_ to perform the conversion.
+    Perceives aromaticity and removes Hydrogen atoms.
+    """
+    rdkitmol = toRDKitMol(mol)
+
+    return Chem.MolToSmarts(rdkitmol)
+
+
+def toSMILES(mol):
+    """
+    Convert a molecular structure to an SMILES string.
+
+    If there is a Nitrogen atom present it uses
+    `OpenBabel <http://openbabel.org/>`_ to perform the conversion,
+    and the SMILES may or may not be canonical.
+
+    Otherwise, it uses `RDKit <http://rdkit.org/>`_ to perform the
+    conversion, so it will be canonical SMILES.
+    While converting to an RDMolecule it will perceive aromaticity
+    and removes Hydrogen atoms.
+    """
+
+    # If we're going to have to check the formula anyway,
+    # we may as well shortcut a few small known molecules.
+    # Dictionary lookups are O(1) so this should be fast:
+    # The dictionary is defined at the top of this file.
+
+    cython.declare(
+        atom=Atom,
+        # obmol=,
+        # rdkitmol=,
+    )
+
+    try:
+        if mol.isRadical():
+            return _known_smiles_radicals[mol.getFormula()]
+        else:
+            return _known_smiles_molecules[mol.getFormula()]
+    except KeyError:
+        # It wasn't in the above list.
+        pass
+    for atom in mol.vertices:
+        if atom.isNitrogen():
+            obmol = toOBMol(mol)
+            try:
+                SMILEwriter = openbabel.OBConversion()
+                SMILEwriter.SetOutFormat('smi')
+                SMILEwriter.SetOptions("i",
+                                       SMILEwriter.OUTOPTIONS)  # turn off isomer and stereochemistry information (the @ signs!)
+            except:
+                pass
+            return SMILEwriter.WriteString(obmol).strip()
+
+    rdkitmol = toRDKitMol(mol, sanitize=False)
+    if not mol.isAromatic():
+        return Chem.MolToSmiles(rdkitmol, kekuleSmiles=True)
+    return Chem.MolToSmiles(rdkitmol)
+
+
 
 def __fromSMILES(mol, smilesstr, backend):
     """Replace the Molecule `mol` with that given by the SMILES `smilesstr`
@@ -109,17 +324,19 @@ def __fromSMILES(mol, smilesstr, backend):
     else:
         raise NotImplementedError('Unrecognized backend for SMILES parsing: {0}'.format(backend))
 
+
 def __fromInChI(mol, inchistr, backend):
     """Replace the Molecule `mol` with that given by the InChI `inchistr`
        using the backend `backend`"""
     if backend.lower() == 'rdkit':
         rdkitmol = Chem.inchi.MolFromInchi(inchistr, removeHs=False)
         mol = fromRDKitMol(mol, rdkitmol)
-        return mol 
+        return mol
     elif backend.lower() == 'openbabel':
         return parse_openbabel(mol, inchistr, 'inchi')
     else:
         raise NotImplementedError('Unrecognized backend for InChI parsing: {0}'.format(backend))
+
 
 def __fromSMARTS(mol, smartsstr, backend):
     """Replace the Molecule `mol` with that given by the SMARTS `smartsstr`
@@ -133,15 +350,16 @@ def __fromSMARTS(mol, smartsstr, backend):
     else:
         raise NotImplementedError('Unrecognized backend for SMARTS parsing: {0}'.format(backend))
 
+
 def __parse(mol, identifier, type_identifier, backend):
     """
     Parses the identifier based on the type of identifier (inchi/smi/sma)
     and the backend used.
-    
+
     First, look up the identifier in a dictionary to see if it can be processed
     this way.
 
-    If not in the dictionary, parse it through the specified backed, 
+    If not in the dictionary, parse it through the specified backed,
     or try all backends.
 
     """
@@ -151,7 +369,7 @@ def __parse(mol, identifier, type_identifier, backend):
             mol.updateAtomTypes()
             return mol
 
-    for _backend in (BACKENDS if backend=='try-all' else [backend]):
+    for _backend in (BACKENDS if backend == 'try-all' else [backend]):
         if type_identifier == 'smi':
             __fromSMILES(mol, identifier, _backend)
         elif type_identifier == 'inchi':
@@ -170,10 +388,11 @@ def __parse(mol, identifier, type_identifier, backend):
     logging.error("Unable to correctly parse %s with backend %s", identifier, backend)
     raise Exception("Couldn't parse {0}".format(identifier))
 
+
 def parse_openbabel(mol, identifier, type_identifier):
     """Converts the identifier to a Molecule using Openbabel."""
     obConversion = openbabel.OBConversion()
-    obConversion.SetInAndOutFormats(type_identifier, "smi")#SetInFormat(identifier) does not exist.
+    obConversion.SetInAndOutFormats(type_identifier, "smi")  # SetInFormat(identifier) does not exist.
     obmol = openbabel.OBMol()
     obConversion.ReadString(obmol, identifier)
     obmol.AddHydrogens()
@@ -199,6 +418,7 @@ def isCorrectlyParsed(mol, identifier):
 
     return all(conditions)
 
+
 def __lookup(mol, identifier, type_identifier):
     """
     Looks up the identifier and parses it the way we think is best.
@@ -223,7 +443,7 @@ def __lookup(mol, identifier, type_identifier):
 
 def fromInChI(mol, inchistr, backend='try-all'):
     """
-    Convert an InChI string `inchistr` to a molecular structure. Uses 
+    Convert an InChI string `inchistr` to a molecular structure. Uses
     a user-specified backend for conversion, currently supporting
     rdkit (default) and openbabel.
     """
@@ -234,7 +454,6 @@ def fromInChI(mol, inchistr, backend='try-all'):
         return __parse(mol, inchistr, 'inchi', backend)
     else:
         return __parse(mol, inchiutil.INCHI_PREFIX + '/' + inchistr, 'inchi', backend)
-
 
 
 def fromAugmentedInChI(mol, aug_inchi):
@@ -265,16 +484,17 @@ def fromAugmentedInChI(mol, aug_inchi):
 
     return mol
 
+
 def fromSMILES(mol, smilesstr, backend='try-all'):
     """
-    Convert a SMILES string `smilesstr` to a molecular structure. Uses 
+    Convert a SMILES string `smilesstr` to a molecular structure. Uses
     a user-specified backend for conversion, currently supporting
     rdkit (default) and openbabel.
     """
     return __parse(mol, smilesstr, 'smi', backend)
 
 
-def fromSMARTS(mol, smartsstr, backend = 'rdkit'):
+def fromSMARTS(mol, smartsstr, backend='rdkit'):
     """
     Convert a SMARTS string `smartsstr` to a molecular structure. Uses
     `RDKit <http://rdkit.org/>`_ to perform the conversion.
